+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,12 @@
 # docs
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - pinch-zoom-native@0.0.10
+
 ## 0.0.10
 
 ### Patch Changes